Add CategoryFilterCollection
